quietly and only if not set

+++ R/default_auth.R
@@ -1,13 +1,16 @@
 default <- function(what) {
 
-  href <- paste0("https://raw.githubusercontent.com/",
-         "boettiger-lab/earthdatalogin/main/inst/.Renviron")
-  renviron <- tempfile("Renviron")
-  download.file(href, renviron)
-  tryCatch(readRenviron(renviron),
-           warning = function(e) NULL)
-  unlink(renviron) # clean up
-  switch(what,
-         user = Sys.getenv("EARTHDATA_USER", "earthaccess"),
-         password = Sys.getenv("EARTHDATA_PASSWORD", "EDL_test1"))
+  if (Sys.getenv("EARTHDATA_USER") == "" ||
+      Sys.getenv("EARTHDATA_PASSWORD") == "") {
+    href <- paste0("https://raw.githubusercontent.com/",
+           "boettiger-lab/earthdatalogin/main/inst/.Renviron")
+    renviron <- tempfile("Renviron")
+    download.file(href, renviron, quiet = TRUE)
+    tryCatch(readRenviron(renviron),
+             warning = function(e) NULL)
+    unlink(renviron) # clean up
+    switch(what,
+           user = Sys.getenv("EARTHDATA_USER", "earthaccess"),
+           password = Sys.getenv("EARTHDATA_PASSWORD", "EDL_test1"))
+    }
 }
